feat: Migrate to minIO and finish up users profile pics
migrated form azurite because it was pain in the ass to use in
production with no tooling for storage explore. Added minio client and
compose setup, on 9000 admin ui is available to explore buckets and
objects. Added mapping of public facing url for profile pictures to user
DTO model, minio should be publicly accessible for this to work since
this is just url straight to that. Added setup of buckets with public
access policies on database schema setup.